# MAINT: Optimize numpy.count_nonzero for int types using SIMD operations