… numpy/core/tests/test_numeric.py