As pointed out in [this issue](https://github.com/numpy/numpy/issues/11569), ```numpy.count_nonzero```, ```numpy.nonzero```, ```numpy.flatnonzero``` are rather slow which could use some optimization. This PR optimizes ```numpy.count_nonzero``` for signed and unsigned 8 bit, 16 bit, 32 bit and 64 bit integers using SIMD operations. This in turn speeds up ```numpy.flatnonzero```, ```numpy.nonzero```, and several other functions that depend on ```numpy.count_nonzero```. 

Below, I have given benchmarks to showcase the speed improvements for the integer types with AVX2.
```python
import numpy as np

np.random.seed(0)
x = np.random.randint(0,2, size=10**6, dtype=np.int8)
%timeit -n 1000 -r 30 np.count_nonzero(x)
# With SIMD optimization: 34.9 µs ± 1.11 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)
# Without SIMD optimization: 1.96 ms ± 86 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)

import numpy as np
np.random.seed(0)
x = np.random.randint(0,2, size=10**6, dtype=np.int16)
%timeit -n 1000 -r 30 np.count_nonzero(x)
# With SIMD optimization: 55.9 µs ± 1.81 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)
# Without SIMD optimization: 1.94 ms ± 64.7 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)

import numpy as np
np.random.seed(0)
x = np.random.randint(0,2, size=10**6, dtype=np.int32)
%timeit -n 1000 -r 30 np.count_nonzero(x)
# With SIMD optimization: 138 µs ± 4.75 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)
# Without SIMD optimization: 1.93 ms ± 31.3 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)

import numpy as np
np.random.seed(0)
x = np.random.randint(0,2, size=10**6, dtype=np.int64)
%timeit -n 1000 -r 30 np.count_nonzero(x)
# With SIMD optimization: 387 µs ± 5.76 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)
# Without SIMD optimization: 2.02 ms ± 18.8 µs per loop (mean ± std. dev. of 30 runs, 1000 loops each)

``` 

I have added few test cases for each of the integer types. Let me know if more is required.