@touqir14 2021-01-22
@gnool , without further investigation, I am speculating the overhead is coming from the use of an iterator and calls to ```get_multi_index``` function in https://github.com/numpy/numpy/blob/master/numpy/core/src/multiarray/item_selection.c#L2555-L2600 whenever a multidimensional array is passed. In contrast, when a single dimensional array is passed, without an iterator the indices are computed using the ```npy_memchr``` function which is quite fast : https://github.com/numpy/numpy/blob/master/numpy/core/src/multiarray/item_selection.c#L2485-L2520