@tylerjereddy 2021-01-29
I make heavy use of `numpy.nonzero()` on computational geometry work--will be great to see some potential speedups there.