@touqir14 2021-02-05
@tylerjereddy , can you point me to some of your use cases of ```nonzero``` ? I think I **might** be able to speed it up for certain cases.