@tylerjereddy 2021-02-06
@touqir14 Finding the indices where an extremely large array of bools is `True` for example.

Maybe something like:

```python

# test_array is a huge 1D array of np.float64
relevant_indices = np.nonzero(test_array > 0.5)

```