@touqir14 2021-02-07
@Qiyu8 , @seiko2plus , I also want to add optimizations to ```PyArray_Nonzero``` for some special input cases(Ndarray is aligned and C contiguous). Should I just keep sending commits in this PR?