@seiko2plus 2021-02-07
@touqir14,
> Should I just keep sending commits in this PR?

let's keep this pull-request only for `count_nonzero`