@touqir14 2021-02-08
@seiko2plus , the overflow possibilities have been taken care of. Please see my last commit to verify. Looks like distinguishing dtypes using ```kind``` and ```elsize``` has taken away the build issues on win32/64 and linux 32bit. There is a failed test issue regarding numpy versioning which however has nothing to do with my commits.