@seiko2plus 2021-02-13
@touqir14, I made some changes in order to reduce the amount of code, it wouldn't affect performance.

> are we all good now? If so, please merge this PR.

yes, I think it's good. I prefer to wait one day more to give a chance to the others to look at the code.